fix: collect target allocation during the style step
The flow now keeps confirmation focused on OCR cleanup and moves
custom target allocation into the style selection step, so users can
pick a preset or adjust sliders in the same place before diagnosis.

Constraint: Keep the target slider markup and styles aligned with the existing confirm-step implementation
Rejected: Leave a skip button on /style | it no longer matches the balanced-target gate
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Preserve the /style rule that custom slider targets can proceed without a preset choice
Tested: COREPACK_ENABLE_AUTO_PIN=0 corepack pnpm@9.15.4 verify
Not-tested: Manual browser walkthrough of the /confirm -> /style flow
